+++ docs/changelog.rst
@@ -7,6 +7,8 @@ Changelog
 Under development
 
 * Updated homepage for extension
+* Adapted extension activationEvents to workspace contains ``conf.py``.
+* Added setting :ref:`activateFiles` to config supported files that language features will be activated when file opened.
 
 0.1.1
 -----